+++ src/libgenesis13/loading/Profile.cpp
@@ -109,7 +109,7 @@ string ProfilePolynom::init(int rank, map<string, string>* arg) {
     string label="";
     map<string, string>::iterator end=arg->end();
     c.resize(5);
-    for (int i=0; i< c.size(); i++) {
+    for (size_t i=0; i< c.size(); i++) {
         c[i]=0;
     }
     if (arg->find("label")!=end) {
@@ -154,7 +154,7 @@ string ProfilePolynom::init(int rank, map<string, string>* arg) {
 double ProfilePolynom::value(double z) {
     double val=0;
     double zsave=1;
-    for (int i=0; i<c.size(); i++) {
+    for (size_t i=0; i<c.size(); i++) {
         val+=c[i]*zsave;
         zsave*=z;
     }
@@ -341,7 +341,7 @@ string ProfileFile::init(int rank, map<string, string>* arg) {
         return "";
     }
     if (isTime) {
-        for (int i=0; i<xdat.size(); i++) {
+        for (size_t i=0; i<xdat.size(); i++) {
             xdat[i]*=LIGHT_SPEED;         // scale time variable to space varial by multiplying the speed of light
         }
     }
@@ -350,7 +350,7 @@ string ProfileFile::init(int rank, map<string, string>* arg) {
         double xmax=xdat[xdat.size()-1];
         reverse(xdat.begin(), xdat.end());
         reverse(ydat.begin(), ydat.end());
-        for (int i=0; i<xdat.size(); i++) {
+        for (size_t i=0; i<xdat.size(); i++) {
             xdat[i]=-xdat[i]+xmin+xmax;    // get the correct time window
         }
     }
